+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ News
 - Fixed #73. Python < 2.6.5 does not support unicode as keyword arguments names.
   [Stepan Kolesnik]
 
+- Fixed #89 remove WSGIWarning: You are not supposed to send a body in a DELETE
+  request because we now have a good reason for that. See http://bit.ly/1tb3yxW
+  [gawel]
+
 - Fixed #92 You can now override TestApp.JSONEncoder to use a custom encoder
   [gawel]
 
@@ -20,6 +24,7 @@ News
   [gawel]
 
 - Fixed #108 cgi.parse_qsl is pending deprecation
+  [gawel]
 
 2.0.14 (2014-01-23)
 -------------------